Remove whenAgreed hoc, add withApplication hoc instead
