+++ pages/apply/household/add-person.tsx
@@ -2,12 +2,12 @@ import { useRouter } from 'next/router';
 import { HeadingOne } from '../../../components/content/headings';
 import Form from '../../../components/form/form';
 import Layout from '../../../components/layout/resident-layout';
-import whenAgreed from '../../../lib/hoc/whenAgreed';
 import { useAppDispatch, useAppSelector } from '../../../lib/store/hooks';
 import { addResidentFromFormData } from '../../../lib/store/otherMembers';
 import { FormData } from '../../../lib/types/form';
 import { FormID, getFormData } from '../../../lib/utils/form-data';
 import { loadApplication } from '../../../lib/store/application';
+import withApplication from '../../../lib/hoc/withApplication';
 
 const AddPersonToApplication = (): JSX.Element => {
   const returnHref = '/apply/household';
@@ -41,4 +41,4 @@ const AddPersonToApplication = (): JSX.Element => {
   );
 };
 
-export default whenAgreed(AddPersonToApplication);
+export default withApplication(AddPersonToApplication);
